+++ packages/docs/src/stories/Button.stories.tsx
@@ -1,4 +1,4 @@
-import { Button, ButtonProps } from '@jordanio-iu/react'
+import { Button, ButtonProps } from '@jordaniodev-iu/react'
 import { Meta, StoryObj } from '@storybook/react'
 import { ArrowRight } from 'phosphor-react'