+++ packages/docs/src/stories/ToolTip.stories.tsx
@@ -1,4 +1,4 @@
-import { Button, Text, ToolTip, ToolTipProps } from '@jordanio-iu/react'
+import { Button, Text, ToolTip, ToolTipProps } from '@jordaniodev-iu/react'
 import { Meta, StoryObj } from '@storybook/react'
 
 export default {