+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jordaniodev-iu/react
 
+## 3.0.0
+
+### Major Changes
+
+- f414053: Rename lib
+
 ## 2.1.0
 
 ### Minor Changes